backend/requirements.txt
fastapi==0.100.0
uvicorn==0.23.2
SQLAlchemy==2.0.20
pydantic==2.5.1
pydantic[email]==2.5.1
python-dotenv==1.0.0
PyJWT==2.8.0
hashids==1.3.1
psycopg2-binary==2.9.7
python-jose==3.3.0
passlib[bcrypt]==1.7.4
python-multipart
